-Navigator cleanup -MaterialCommunityIcon in BottomTabNavigator.js
