+++ package.json
@@ -26,6 +26,7 @@
     "react-native-reanimated": "~2.12.0",
     "react-native-safe-area-context": "4.4.1",
     "react-native-screens": "~3.18.0",
+    "react-native-vector-icons": "^9.2.0",
     "react-native-web": "~0.18.9"
   },
   "devDependencies": {
